added br in manifest file
