+++ app/src/main/AndroidManifest.xml
@@ -18,7 +18,11 @@
                 <category android:name="android.intent.category.LAUNCHER"/>
             </intent-filter>
         </activity>
-
+        <receiver android:name=".NetworkConnectivityChangeReceiver">
+            <intent-filter> 
+                <action android:name="android.net.conn.CONNECTIVITY_CHANGE" /> 
+            </intent-filter>
+        </receiver>
         <service
             android:name=".MQTTService"
             android:enabled="true"
